+++ src/main/java/gg/mineral/server/network/packet/play/clientbound/DestroyEntitiesPacket.java
@@ -20,7 +20,7 @@ public class DestroyEntitiesPacket implements Packet.OUTGOING {
     }
 
     @Override
-    public int getId() {
+    public byte getId() {
         return 0x13;
     }
