+++ src/main/java/gg/mineral/server/network/packet/play/clientbound/EntityLookPacket.java
@@ -22,7 +22,7 @@ public class EntityLookPacket implements Packet.OUTGOING {
     }
 
     @Override
-    public int getId() {
+    public byte getId() {
         return 0x16;
     }
