+++ src/main/java/gg/mineral/server/network/packet/play/clientbound/SpawnGlobalEntityPacket.java
@@ -27,7 +27,7 @@ public class SpawnGlobalEntityPacket implements Packet.OUTGOING {
     }
 
     @Override
-    public int getId() {
+    public byte getId() {
         return 0x2C;
     }
