+++ src/main/java/gg/mineral/server/network/protocol/ProtocolState.java
@@ -40,7 +40,7 @@ public class ProtocolState {
 
         @Override
         public void registerPackets() {
-            put(0x00, () -> new HandshakePacket());
+            put((byte) 0x00, () -> new HandshakePacket());
         }
 
     };
@@ -49,8 +49,8 @@ public class ProtocolState {
 
         @Override
         public void registerPackets() {
-            put(0x00, () -> new LoginStartPacket());
-            put(0x01, () -> new EncryptionKeyResponsePacket());
+            put((byte) 0x00, () -> new LoginStartPacket());
+            put((byte) 0x01, () -> new EncryptionKeyResponsePacket());
         }
 
     };
@@ -59,30 +59,30 @@ public class ProtocolState {
 
         @Override
         public void registerPackets() {
-            put(0x0A, () -> new AnimationPacket());
-            put(0x01, () -> new ChatMessagePacket());
-            put(0x09, () -> new HeldItemChangePacket());
-            put(0x00, () -> new KeepAlivePacket());
-            put(0x0E, () -> new ClickWindowPacket());
-            put(0x15, () -> new ClientSettingsPacket());
-            put(0x16, () -> new ClientStatusPacket());
-            put(0x0D, () -> new CloseWindowPacket());
-            put(0x0F, () -> new ConfirmTransactionPacket());
-            put(0x10, () -> new CreativeInventoryActionPacket());
-            put(0x11, () -> new EnchantItemPacket());
-            put(0x0B, () -> new EntityActionPacket());
-            put(0x13, () -> new PlayerAbilitiesPacket());
-            put(0x08, () -> new PlayerBlockPlacementPacket());
-            put(0x07, () -> new PlayerDiggingPacket());
-            put(0x05, () -> new PlayerLookPacket());
-            put(0x03, () -> new PlayerPacket());
-            put(0x06, () -> new PlayerPositionAndLookPacket());
-            put(0x04, () -> new PlayerPositionPacket());
-            put(0x17, () -> new PluginMessagePacket());
-            put(0x0C, () -> new SteerVehiclePacket());
-            put(0x14, () -> new TabCompletePacket());
-            put(0x12, () -> new UpdateSignPacket());
-            put(0x02, () -> new UseEntityPacket());
+            put((byte) 0x0A, () -> new AnimationPacket());
+            put((byte) 0x01, () -> new ChatMessagePacket());
+            put((byte) 0x09, () -> new HeldItemChangePacket());
+            put((byte) 0x00, () -> new KeepAlivePacket());
+            put((byte) 0x0E, () -> new ClickWindowPacket());
+            put((byte) 0x15, () -> new ClientSettingsPacket());
+            put((byte) 0x16, () -> new ClientStatusPacket());
+            put((byte) 0x0D, () -> new CloseWindowPacket());
+            put((byte) 0x0F, () -> new ConfirmTransactionPacket());
+            put((byte) 0x10, () -> new CreativeInventoryActionPacket());
+            put((byte) 0x11, () -> new EnchantItemPacket());
+            put((byte) 0x0B, () -> new EntityActionPacket());
+            put((byte) 0x13, () -> new PlayerAbilitiesPacket());
+            put((byte) 0x08, () -> new PlayerBlockPlacementPacket());
+            put((byte) 0x07, () -> new PlayerDiggingPacket());
+            put((byte) 0x05, () -> new PlayerLookPacket());
+            put((byte) 0x03, () -> new PlayerPacket());
+            put((byte) 0x06, () -> new PlayerPositionAndLookPacket());
+            put((byte) 0x04, () -> new PlayerPositionPacket());
+            put((byte) 0x17, () -> new PluginMessagePacket());
+            put((byte) 0x0C, () -> new SteerVehiclePacket());
+            put((byte) 0x14, () -> new TabCompletePacket());
+            put((byte) 0x12, () -> new UpdateSignPacket());
+            put((byte) 0x02, () -> new UseEntityPacket());
         }
 
     };
@@ -91,8 +91,8 @@ public class ProtocolState {
 
         @Override
         public void registerPackets() {
-            put(0x00, () -> new RequestPacket());
-            put(0x01, () -> new PingPacket());
+            put((byte) 0x00, () -> new RequestPacket());
+            put((byte) 0x01, () -> new PingPacket());
         }
 
     };
