+++ src/main/java/gg/mineral/server/network/protocol/ProtocolVersion.java
@@ -1,5 +1,5 @@
 package gg.mineral.server.network.protocol;
 
 public class ProtocolVersion {
-    public static final int V1_7_6 = 5;
+    public static final byte V1_7_6 = 5;
 }
